Move ComputeSSB192TheoreticalSubs into pipeline_utils and setup v2
